Create Subspace class
